fixed type error

+++ eufs_rqt/src/eufs_rqt/EUFSRobotSteeringGUI.py
@@ -351,7 +351,7 @@ class EUFSRobotSteeringGUI(Plugin):
 
     def _on_max_linear_changed(self, value):
         self._widget.linear_slider.setMaximum(
-            value * EUFSRobotSteeringGUI.slider_factor)
+            int(value * EUFSRobotSteeringGUI.slider_factor))
 
     def _on_min_linear_changed(self, value):
         self._widget.linear_slider.setMinimum(
